Refactor Service class
